Convert labels to long type for index tensor

+++ notebooks/pipe.py
@@ -422,7 +422,7 @@ if args.train:
                     outputs = model(inputs)
                     labels = labels.to(outputs.device)  # Move labels tensor to the same device as outputs
                     labels = labels.view(-1, 1)  # Reshape labels tensor to match the shape of outputs tensor
-                    labels = labels.float()  # Convert labels to index tensor
+                    labels = labels.long()  # Convert labels to index tensor
                     labels = F.one_hot(labels, num_classes=2)  # Convert labels to one-hot binary
                     logger.info(f"Outputs: {outputs}")
                     logger.info(f"Labels: {labels}")
